Specifically reorient RAO images
Using only pinhole camera assumptions, LAO and RAO images would appear flipped relative to each other. However, they appear to almost always have the same orientation, perhaps to reduce mental burden on the clinicians. This same logic underlies why AP and PA images appear in the same orientation as well. Therefore, if an RAO image is given in P/F orientation, we flip it to A/F, which is how it would appear from a rendering.